mithril-os/render_template/pyproject.toml
[tool.poetry]
name = "render-template"
version = "0.1.0"
description = ""
authors = ["Corentin Lauverjat <corentin.lauverjat@mithrilsecurity.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
render_template = "render_template.__main__:main"